French cup structure
