+++ python/evolutek/lib/robot/actions/build_cakes.py
@@ -1,0 +1,7 @@
+from evolutek.lib.robot.robot_actions_imports import *
+
+
+@if_enabled
+@async_task
+def build_cakes(self):
+    return check_status()
