+++ python/evolutek/lib/robot/actions/place_cherry.py
@@ -1,0 +1,6 @@
+from evolutek.lib.robot.robot_actions_imports import *
+
+@if_enabled
+@async_task
+def fill_n_cherries(self, n):
+    return check_status()
