+++ python/evolutek/lib/robot/robot_actions.py
@@ -1,5 +1,7 @@
-from evolutek.lib.robot.actions.goto_random import goto_random
-from evolutek.lib.robot.actions.go_grab import roam_stacks, roam_zones, go_grab_one_stack, go_drop_all, stack_and_grab, grab_first_stacks, back_to_base, drop_1_stack_go_back
-from evolutek.lib.robot.actions.empty_cherries import empty_n_cherries, empty_all_cherries
-from evolutek.lib.robot.actions.fill_cherries import fill_n_cherries, fill_all_cherries, set_cherry_count
-from evolutek.lib.robot.actions.vaccum_cherry import vacuum_10_cherry_right, vacuum_10_cherry_left
+from evolutek.lib.robot.actions.stack_and_grab import stack_and_grab
+from evolutek.lib.robot.actions.drop import drop_n, drop_all
+from evolutek.lib.robot.actions.suck_rack import suck_rack
+from evolutek.lib.robot.actions.build_cakes import build_cakes
+from evolutek.lib.robot.actions.shoot_cherries import shoot_n_cherries, shoot_all_cherries
+from evolutek.lib.robot.actions.place_cherry import place_cherry
+from evolutek.lib.robot.actions.fill_cherries import fill_n_cherries, set_cherry_count
