+++ python/evolutek/lib/robot/robot_actions_imports.py
@@ -9,3 +9,9 @@ from evolutek.lib.map.point import Point
 
 from time import sleep
 from math import pi
+
+def check_status(*args, score=0):
+    for stat in args:
+        if stat != RobotStatus.Done:
+            return RobotStatus.return_status(RobotStatus.Failed)
+    return RobotStatus.return_status(RobotStatus.Done, score=score)
